Refactor separate music canvas, navbar from page
